+++ changelog.md
@@ -10,6 +10,7 @@
 - Improved tests
 - Moved all data getting API endpoints to a single file
 - Cleaning
+- Added CSRF token!
 
 ## Version 0.1.0 (2023/09/01)